@@ -2,7 +2,7 @@
 name = "trust-graph"
 version = "0.3.0"
 authors = ["Fluence Labs"]
-edition = "2018"
+edition = "2021"
 description = "trust graph"
 license = "Apache-2.0"
 repository = "https://github.com/fluencelabs/trust-graph"
