@@ -2,7 +2,7 @@
 name = "fluence-keypair"
 version = "0.9.0"
 authors = ["Fluence Labs"]
-edition = "2018"
+edition = "2021"
 description = "identity"
 license = "Apache-2.0"
 repository = "https://github.com/fluencelabs/trust-graph"
