@@ -2,7 +2,7 @@
 name = "trust-graph-wasm"
 version = "0.3.1"
 authors = ["Fluence Labs"]
-edition = "2018"
+edition = "2021"
 description = "trust graph wasm"
 license = "Apache-2.0"
 publish = false
